% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-quantiles.R
\name{adjust_dropout}
\alias{adjust_dropout}
\title{Inflate a sample size for anticipated dropout}
\usage{
adjust_dropout(n, dropout)
}
\arguments{
\item{n}{Evaluable sample size (single positive integer).}

\item{dropout}{Dropout proportion in [0, 1).}
}
\description{
Returns \code{ceiling(n / (1 - dropout))}.
}
\references{
Chow, S.C., Shao, J., Wang, H. (2008). \emph{Sample Size Calculations
in Clinical Research}, 2nd ed. Chapman & Hall/CRC. Section 1.2.5.
}
\keyword{internal}
