% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-result.R
\name{print.clinss_result}
\alias{print.clinss_result}
\title{Print a clinss_result object}
\usage{
\method{print}{clinss_result}(x, digits = 4, ...)
}
\arguments{
\item{x}{A \code{clinss_result} object.}

\item{digits}{Number of significant digits for the achieved power.}

\item{...}{Ignored.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a clinss_result object
}
